Condense a reminder list into its own class
Since we will have several independently scrolling lists, it makes sense to create a consolidated ReminderList view. The ReminderStack view is not used at the moment due to compiler difficulties, but it would essentially create a horizontal scrolling view that is two reminders tall when the list of reminders is long enough (for faster scrolling).