Add SharedBool (for now)
Created a wrapper for Bool that allows us to generate as many bindings to it as we want, right out of the gate. This saves a lot of duplicate boilerplate code. However, as of now, I have managed to refactor my initial use case for SharedBool OUT of the code entirely, so we will see if SharedBool even ends up being necessary. (It may not be.)