Rethink how to present card views
I was initially going to make it so that we use UIKit code to handle the transition between the home screen and card views, but I decided to give using SwiftUI alone another shot. What we do here is create a ZStack — one with the background in the back and the primary content up front. The primary content is interchangeable; i.e., it can be either the main screen or an individual card, depending on the state. When the user taps on a card, we store their selection and the position of the card on-screen, then clone the view to the "card stage." Then the primary stage disappears, the duplicated card expands with the UI elements appropriate for the expanded card, and the user can go about their business. Much simpler, I think, than it would have been to implement all of this in UIKit and then try to wrangle it into SwiftUI.