Basic subviews, Group and OnExpand
'Group' is just boilerplate, entirely nonfunctional except as a syntax sweetener.
'OnExpand' allows us to write the expanded / collapsed versions of a user interface without using so many if/elses that the Swift compiler has a heart attack. Using the input parameter provided by the 'Card' class (similar to how we get a local variable out of GeometryReader), we get a binding that controls all OnExpand blocks simultaneously:

Card { expanded in
    OnExpand(controlledBy: expanded) {
        Text("Shows when expanded")
    }
    OnExpand(controlledBy: expanded, showWhenCollapsedInstead: true) {
        Text("Shows when collapsed")
    }
}

Unfortunately, we cannot WRITE to 'expanded,' which means we cannot control how to close the card from within its contents. This is an acceptable limitation for now, since I want users to collapse cards by tapping outside their boundary rectangles; however, it may be necessary to look into allowing that functionality at a later time.